Add utility for converting an array of currentTimes to canvas elements.
Because:
- Previous implementation resulted in visually choppy scrubbing
- Moving functionality to utility promotes reusable code.

Instead of an array of <canvas> elements that are swapped out and
redrawn, there is a single, persistent canvas displayed. With each frame
update, the contents of a corresponding  offscreen canvas element is
copied to the persistent canvas using drawImage(). The result is a smooth,
seamless transition between frames when scrubbed.